pyproject.toml
[project]
name = "stromalytix"
version = "0.1.0"
description = "BioSim Copilot - Tissue engineering decision intelligence platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "langchain",
    "langchain-anthropic",
    "langchain-openai",
    "langchain-community",
    "chromadb",
    "biopython",
    "plotly",
    "pandas",
    "pydantic",
    "python-dotenv",
    "openai",
    "tiktoken",
    "anthropic>=0.84.0",
    "fpdf2>=2.8.7",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "scikit-fem>=12.0.1",
    "networkx>=3.6.1",
    "scikit-learn>=1.8.0",
    "kaleido>=1.2.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.22",
    "aiofiles>=25.1.0",
    "reportlab>=4.4.10",
    "trimesh>=4.0.0",
    "scikit-image>=0.24.0",
    "scipy>=1.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
